% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{ConnectionError}
\alias{ConnectionError}
\alias{ConnectionRefusedError}
\alias{ConnectionAbortedError}
\title{Connection Errors}
\usage{
ConnectionError(message = "Connection error", ..., class = NULL, call = NULL)

ConnectionRefusedError(
  message = "Connection refused by peer",
  ...,
  class = NULL,
  call = NULL
)

ConnectionAbortedError(
  message = "Connection aborted by peer",
  ...,
  class = NULL,
  call = NULL
)
}
\description{
Connection Errors

Raised when a connection attempt is refused by the peer. Corresponds to errno ECONNREFUSED.

Rraised when a connection attempt is aborted by the peer. Corresponds to errno ECONNABORTED.
}
