% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{DirNotEmptyError}
\alias{DirNotEmptyError}
\title{Title}
\usage{
DirNotEmptyError(
  message = sprintf("directory '\%s' does not exist.", dir),
  ...,
  dir,
  class = NULL,
  call = NULL
)
}
\arguments{
\item{call}{}
}
\value{

}
\description{
Title
}
