% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{handle_http_error}
\alias{handle_http_error}
\title{Handle HttpErrors in plumber APIs}
\usage{
handle_http_error(req, res, err)
}
\arguments{
\item{req}{a plumber Request object}

\item{res}{a plumber Response object}

\item{err}{the error}
}
\value{
an \link{http_error_body} list object
}
\description{
Handle HttpErrors in plumber APIs
}
\seealso{
\itemize{
\item https://github.com/rstudio/plumber/pull/507#issuecomment-675693897
\item https://www.rplumber.io/
}
}
