% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{http_error_body}
\alias{http_error_body}
\alias{as_http_error_body}
\alias{as_http_error_body.default}
\title{Error Response Body}
\usage{
http_error_body(message, ..., class = class)

as_http_error_body(e)

\method{as_http_error_body}{default}(e)
}
\arguments{
\item{message}{\code{character} scalar. an error message}

\item{...}{arbitrary \R obejcts, as long as plumber knows how to serialize
them}

\item{e}{a \code{condition} S3 object (usually an \code{error})}

\item{type}{\code{character} scalar. Type of the error (usually \code{class(e)[[1]]})}

\item{error}{\code{character} an \link[base:conditions]{base::errorCondition} object}
}
\value{
A \code{list} of subclass \code{http_error_body} with elements \code{error}
(the http), \code{type},
and \code{message}
}
\description{
Error Response Body

\code{as_http_error_body()} converts \R condition objects to lists that plumber knows
how to serialze to.
}
\seealso{
\link{errors}
}
