% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{HttpError}
\alias{HttpError}
\alias{HttpServerError}
\alias{HttpInternalServerError}
\alias{HttpServiceUnavailableError}
\alias{HttpClientError}
\alias{HttpBadRequestError}
\title{HTTP errors}
\usage{
HttpError(message, status, ..., headers = NULL, class = NULL, call = NULL)

HttpServerError(
  message,
  status,
  ...,
  headers = NULL,
  class = NULL,
  call = NULL
)

HttpInternalServerError(
  message = "Internal Server Error",
  ...,
  headers = NULL,
  class = NULL,
  call = NULL
)

HttpServiceUnavailableError(
  message = "Service Unavailable",
  ...,
  headers = NULL,
  class = NULL,
  call = NULL,
  retry_after = NULL
)

HttpClientError(
  message,
  status,
  ...,
  headers = NULL,
  class = NULL,
  call = NULL
)

HttpBadRequestError(
  message = "Bad Request",
  ...,
  headers = headers,
  class = NULL,
  call = NULL
)
}
\description{
Special error classes for HTTP. See also [as_error_body] for turning R errors
into lists that can be returned by plumber APIs
}
\seealso{
https://tools.ietf.org/html/rfc7231#section-6
}
