% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{IsNotDirError}
\alias{IsNotDirError}
\title{Title}
\usage{
IsNotDirError(
  message = sprintf("'\%s' is not a directory.", dir),
  ...,
  dir,
  class = NULL,
  call = NULL
)
}
\arguments{
\item{call}{}
}
\value{

}
\description{
Title
}
