% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{NotImplementedError}
\alias{NotImplementedError}
\title{Title}
\usage{
NotImplementedError(
  message = sprintf("functionality is not yet implemented", dir),
  ...,
  dir,
  class = NULL,
  call = NULL
)
}
\arguments{
\item{call}{}
}
\value{

}
\description{
Title
}
