% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{error_body}
\alias{error_body}
\alias{as_error_body}
\alias{as_error_body.default}
\title{Error Response Body}
\usage{
error_body(type, message, error = "500 - Internal Server Error", ...)

as_error_body(e)

\method{as_error_body}{default}(e)
}
\arguments{
\item{type}{`character` scalar. Type of the error (usually `class(e)[[1]]`)}

\item{message}{`character` scalar. an error message}

\item{error}{`character` scalar. The http error as character scalar}

\item{...}{arbitrary \R obejcts, as long as plumber knows how to serialize
them}

\item{e}{a `condition` S3 object (usually an `error`)}
}
\value{
A `list` of subclass `error_body` with elements `error`
  (the http), `type`,
  and `message`
}
\description{
Error Response Body

`as_error_body()` converts \R condition objects to lists that plumber knows
how to serialze to.
}
\seealso{
[errors]
}
