% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{HttpError}
\alias{HttpError}
\alias{HttpServerError}
\alias{HttpInternalServerError}
\alias{HttpServiceUnavailableError}
\alias{HttpClientError}
\alias{HttpBadRequestError}
\alias{HttpNotFoundError}
\alias{HttpConflictError}
\title{HTTP errors}
\usage{
HttpError(message, status, ..., headers = NULL, class = NULL, call = NULL)

HttpServerError(
  message,
  status,
  ...,
  headers = NULL,
  class = NULL,
  call = NULL
)

HttpInternalServerError(
  message = "Internal Server Error",
  ...,
  headers = NULL,
  class = NULL,
  call = NULL
)

HttpServiceUnavailableError(
  message = "Service Unavailable",
  ...,
  headers = NULL,
  class = NULL,
  call = NULL,
  retry_after = NULL
)

HttpClientError(
  message,
  status,
  ...,
  headers = NULL,
  class = NULL,
  call = NULL
)

HttpBadRequestError(
  message = "Bad Request",
  ...,
  headers = NULL,
  class = NULL,
  call = NULL
)

HttpNotFoundError(
  message = "Not Found",
  ...,
  headers = NULL,
  class = NULL,
  call = NULL
)

HttpConflictError(
  message = "Conflict",
  ...,
  headers = NULL,
  class = NULL,
  call = NULL
)
}
\value{
an \verb{error condition} object with the subclasses
\code{HttpError}, \code{http_error} and \verb{http_<status>}. The latter two are for
compatability with \code{\link[httr:http_condition]{httr::http_condition()}} objects. The error conditions
reflect the hirarchy, e.g. \code{HttpServiceUnavailableError} inherits from
\code{HttpServerError}, \code{HttpError}, \code{http_500}, \code{http_error}.
}
\description{
Special error classes for HTTP. See also \link{as_http_error_body} for turning R errors
into lists that can be returned by plumber APIs.
}
\examples{
print(class(HttpServiceUnavailableError()))

tryCatch(
  stop(HttpServiceUnavailableError()),
  HttpServiceUnavailableError = function(e) print(e)
)

tryCatch(
  stop(HttpServiceUnavailableError()),
  http_503 = function(e) print(e)
)
}
\seealso{
https://tools.ietf.org/html/rfc7231#section-6
}
