% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{HttpError}
\alias{HttpError}
\alias{HttpServiceUnavailableError}
\alias{HttpBadRequestError}
\title{HTTP errors}
\usage{
HttpError(message, http_status, ..., class = NULL, call = NULL)

HttpServiceUnavailableError(message = "Service unavailable", ...)

HttpBadRequestError(
  message = "HTTP 400 (Bad Request)",
  http_status = 400,
  ...,
  class = NULL,
  call = NULL
)
}
\description{
Special error classes for HTTP. See also [as_error_body] for turning R errors
into lists that can be returned by plumber apis.
}
